+++ src/i18next.js
@@ -15,7 +15,7 @@ import * as compat from './compatibility/v1';
 class I18n extends EventEmitter {
   constructor(options = {}, callback) {
     super();
-    this.options = options;
+    this.options = transformOptions(options);
     this.services = {};
     this.logger = baseLogger;
     this.modules = {};
@@ -28,6 +28,8 @@ class I18n extends EventEmitter {
       callback = options;
       options = {};
     }
+    if (!options) options = {};
+
     if (options.compatibilityAPI === 'v1') {
       this.options = {...getDefaults(), ...transformOptions(compat.convertAPIOptions(options)), ...{}};
     } else if (options.compatibilityJSON === 'v1') {
@@ -61,7 +63,7 @@ class I18n extends EventEmitter {
         s.cacheConnector.save();
       });
       s.languageUtils = lu;
-      s.pluralResolver = new PluralResolver(lu, {prepend: '_', compatibilityJSON:  this.options.compatibilityJSON});
+      s.pluralResolver = new PluralResolver(lu, {prepend: this.options.pluralSeparator, compatibilityJSON:  this.options.compatibilityJSON});
       s.interpolator = new Interpolator(this.options);
 
       s.backendConnector = new BackendConnector(createClassOnDemand(this.modules.backend), s.resourceStore, s, this.options);
@@ -93,7 +95,7 @@ class I18n extends EventEmitter {
     }
 
     // append api
-    const storeApi = ['addResource', 'addResources', 'addResourceBundle', 'removeResourceBundle', 'hasResourceBundle', 'getResourceBundle'];
+    const storeApi = ['getResource', 'addResource', 'addResources', 'addResourceBundle', 'removeResourceBundle', 'hasResourceBundle', 'getResourceBundle'];
     storeApi.forEach(fcName => {
       this[fcName] = function() { return this.store[fcName].apply(this.store, arguments); };
     });
@@ -113,6 +115,8 @@ class I18n extends EventEmitter {
     if (!callback) callback = () => {};
 
     if (!this.options.resources) {
+      if (this.language && this.language.toLowerCase() === 'cimode') return callback(); // avoid loading resources for cimode
+
       let toLoad = [];
 
       let append = lng => {